+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/menu/advanced_item_collector/AdvancedItemCollectorScreen.kt
@@ -111,7 +111,7 @@ class AdvancedItemCollectorScreen(
 
 
 		this.buttonRaiseZ = Button.Builder(
-			Component.literal("-X")
+			Component.literal("-Z")
 		) {
 			val packet = ClientClickedMenuButton(AdvancedItemCollectorMenu.RAISE_Z_BUTTON_ID)
 			packet.messageServer()