+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/menu/void_stone/VoidStoneScreen.kt
@@ -13,4 +13,8 @@ class VoidStoneScreen(
 
 	override val background = ScreenTextures.Background.VoidStone
 
+	override fun baseInit() {
+		this.inventoryLabelY = this.imageHeight - 94
+	}
+
 }