+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/entity/SpiritEntity.kt
@@ -128,9 +128,11 @@ class SpiritEntity(
 	}
 
 	private fun ageAndDie() {
+		if (this.hasCustomName()) return
 		age++
+
 		if (age > ServerConfig.CONFIG.spiritMaxAge.get()) {
-//			kill()
+			kill()
 		}
 	}