+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 
 - Fixed the Advanced Item Collector's screen saying `-X` instead of `+Z`
+- Fixed the Void Stone's screen having the inventory label in the wrong place
 
 # 1.2.0