+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Allowed enchanted items to damage Spirits
 - Added some missing localization
 - Spirits can now be damaged by damage types with the `#minecraft:bypasses_invulnerability` tag
+- Fixed a crash when using a mod that adds more enum values to DyeColor (#7)
 
 # 1.2.0